backend/requirements_tests.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-localflavor==5.0
djangorestframework==3.16.0
ecdsa==0.19.1
factory_boy==3.3.3
Faker==37.5.3
idna==3.10
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-stdnum==2.1
qrcode==8.2
redis==6.2.0
reportlab==4.4.3
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
